Move utility classes to own folder.
