+++ src/Utilities/Config.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Yajra\DataTables;
+namespace Yajra\DataTables\Utilities;
 
 use Illuminate\Contracts\Config\Repository;
