+++ src/Utilities/Helper.php
@@ -1,17 +1,11 @@
 <?php
 
-namespace Yajra\DataTables;
+namespace Yajra\DataTables\Utilities;
 
 use DateTime;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Support\Str;
 
-/**
- * Class Helper.
- *
- * @package Yajra\Datatables
- * @author  Arjay Angeles <aqangeles@gmail.com>
- */
 class Helper
 {
     /**
